Add the 'tidy blog' template and begin switching to that for our css
Remember to NOT set TEMPLATE_DIRS in settings.py because we set it in
the settings .ini file that is used by settings.py

Add some stuff for asforums, including a script to populate it with
some test data.

Make it so that we put "MEDIA_URL" in the template context so that
our template can use this to determine where to load media files from.



git-svn-id: file:///Volumes/extra1/backups/locker/repository/as-subversion/wowzer/head@1152 b85eb097-2c14-0410-a966-eec4f2146687